flash/nor/stm32f2x: fix OTP write protection
This fixes some bugs in the STM32F2X driver's handling of the OTP lock
area, which prevented locking the OTP flash.

This writes to the OTP lock area as a series of single-byte writes, mostly
for the sake of simplicity. A more sophisticated implementation could read
back the entire lock area, check if it needs to be changed, and write it
back as doublewords.

While we're in here, I moved the OTP-enabled safety check from
stm32x_write_block to stm32x_write, so single-byte writes can't evade the
safety check.

Change-Id: Id5d7e59a416f2d2789bbd74d7e320a1ff0c5d3d2
Signed-off-by: Will Rigby <will@breakfastny.com>